+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 --------------------------------------------------------------------------------
 
 
-## Unreleased
+## 1.8.1 - Unreleased
 - Fix undefined background color set to map backgroundColor
+- Add track position 'along'. Show track as a single slot centered on the backbone.
+- Change track position 'both' to 'around'. 'both' is depcreated and will be removed in v1.9
 
 
 ## 1.8.0 - 2025-08-29